refactor: move AsciiSet related code to ascii_set.rs
Makes it a little easier to just look at the code related to the set rather than encoding
